move helper funcs to common
